+++ src/Router/router.ts
@@ -1,0 +1,87 @@
+import type IRouter from "./router.interface";
+
+class Router implements IRouter {
+  private routes: Route[];
+
+  constructor() {
+    this.routes = [];
+  }
+
+  addRoute(url: string, method: METHODS[], callback: RouteCallback): void {
+    this.routes.push({
+      url,
+      method,
+      callback,
+    });
+  }
+
+  addRoutes(routes: [Route]) {
+    routes.forEach((route) => {
+      this.addRoute(route.url, route.method, route.callback);
+    });
+  }
+
+  getRoute(url: string | undefined, method: string | undefined): Route | null {
+    if (!url || !method) {
+      return null;
+    }
+
+    let route: Route | null = null;
+
+    this.routes.map((rt: Route) => {
+      const routeParts = rt.url.split("/").filter(Boolean);
+      const urlParts = url.split("/").filter(Boolean);
+
+      if (routeParts.length === urlParts.length) {
+        const match = routeParts.every((part, index) => {
+          return part.startsWith(":") || part === urlParts[index];
+        });
+
+        if (match) {
+          route = rt;
+        }
+      }
+    });
+
+    return route;
+  }
+
+  getPathVariables(
+    reqUrl: string | undefined,
+    routeUrl: string
+  ): Record<string, string> {
+    if (!reqUrl) return {};
+
+    const routeParts = routeUrl.split("/").filter(Boolean);
+    const urlParts = reqUrl.split("?")[0].split("/").filter(Boolean);
+
+    const pathVariables: Record<string, string> = {};
+
+    routeParts.forEach((part: string, index: number) => {
+      if (part.startsWith(":")) {
+        const key = part.slice(1);
+        pathVariables[key] = urlParts[index];
+      }
+    });
+
+    return pathVariables;
+  }
+
+  getqueryParams(url: string | undefined): Record<string, string> {
+    if (!url) return {};
+
+    const queryParams: Record<string, string> = {};
+
+    const queryString = url.split("?")[1];
+    if (!queryString) return queryParams;
+
+    const pairs = queryString.split("&");
+    pairs.forEach((pair: string) => {
+      const [key, value] = pair.split("=");
+      queryParams[decodeURIComponent(key)] = decodeURIComponent(value);
+    });
+
+    return queryParams;
+  }
+}
+export default Router;