+++ src/index.ts
@@ -1,0 +1,5 @@
+import Server from "./Server";
+
+const Simpler = Server
+
+export default Simpler